+++ Common/src/main/java/net/potionstudios/biomeswevegone/BiomesWeveGone.java
@@ -8,6 +8,7 @@ import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.ai.goal.target.NearestAttackableTargetGoal;
+import net.potionstudios.biomeswevegone.config.configs.BWGMobSpawnConfig;
 import net.potionstudios.biomeswevegone.sounds.BWGSounds;
 import net.potionstudios.biomeswevegone.compat.vanilla.dispenser.BWGDispenseItemBehavior;
 import net.potionstudios.biomeswevegone.tags.BWGEntityTypeTags;
@@ -84,7 +85,7 @@ public class BiomesWeveGone {
      * Ran later in the initialization process to setup common things.
      */
     public static void commonSetup() {
-
+        BWGMobSpawnConfig.reload();
     }
 
     /**