+++ Common/src/main/java/net/potionstudios/biomeswevegone/commands/BWGReloadCommand.java
@@ -4,7 +4,6 @@ import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import net.minecraft.ChatFormatting;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.network.chat.Component;
-import net.potionstudios.biomeswevegone.config.ConfigLoader;
 import net.potionstudios.biomeswevegone.config.configs.BWGMobSpawnConfig;
 
 class BWGReloadCommand {
@@ -13,7 +12,7 @@ class BWGReloadCommand {
 		LiteralArgumentBuilder<CommandSourceStack> reload = LiteralArgumentBuilder.literal("reload");
 		reload.requires(commandSourceStack -> commandSourceStack.hasPermission(2));
 		reload.executes(context -> {
-			BWGMobSpawnConfig.INSTANCE = ConfigLoader.loadConfig(BWGMobSpawnConfig.class, "spawn").spawn;
+			BWGMobSpawnConfig.reload();
 			context.getSource().sendSuccess(() -> Component.translatable("biomeswevegone.commands.reload.success").withStyle(ChatFormatting.GREEN), true);
 			return 1;
 		});
@@ -21,7 +20,7 @@ class BWGReloadCommand {
 		LiteralArgumentBuilder<CommandSourceStack> reloadSpawn = LiteralArgumentBuilder.literal("spawn");
 		reloadSpawn.requires(commandSourceStack -> commandSourceStack.hasPermission(2));
 		reloadSpawn.executes(context -> {
-			BWGMobSpawnConfig.INSTANCE = ConfigLoader.loadConfig(BWGMobSpawnConfig.class, "spawn").spawn;
+			BWGMobSpawnConfig.reload();
 			context.getSource().sendSuccess(() -> Component.translatable("biomeswevegone.commands.reload.spawn.success").withStyle(ChatFormatting.GREEN), true);
 			return 1;
 		});