+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.6.2
 - Fix White Puffball not growing over time
+- Fix Misc and Mob Spawn Configs not generating sometimes
 
 # 1.6.1
 - Add Wreaths to Compostable Blocks